OD-18978 feat: replace long values for payment amounts with bigdecimal

+++ server/src/main/groovy/ish/oncourse/server/services/chargebee/property/ChargebeePropertyProcessor.groovy
@@ -32,6 +32,6 @@ abstract class ChargebeePropertyProcessor {
         return SQL_DATE_FORMAT.format(endDate)
     }
 
-    abstract Long getValue(DataSource dataSource)
+    abstract BigDecimal getValue(DataSource dataSource)
     abstract ChargebeePropertyType getType()
 }
